@@ -1,20 +1,20 @@
 name: myportfolio
 description: A new Flutter project.
 
-publish_to: 'none' 
+publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
-  url_launcher: ^5.7.10
-  font_awesome_flutter: ^8.11.0
-  grouped_list: ^3.7.1
-  transparent_image: ^1.0.0
+  cupertino_icons: ^1.0.5
+  url_launcher: ^6.1.11
+  font_awesome_flutter: ^10.4.0
+  grouped_list: ^5.1.2
+  transparent_image: ^2.0.1
 
 dev_dependencies:
   flutter_test:
@@ -51,6 +51,6 @@ flutter:
     - family: Pacifico
       fonts:
         - asset: fonts/Pacifico-Regular.ttf
-  
+
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
